Partial implementation of searching algorithm
-Robot localizes using us sensor
-Robot moves to start of path, and moves along search path, checking
 for object
- Object determination is implemented, no action is implemented

+++ src/ev3ObjectDetector/ObstacleAvoider.java
@@ -58,7 +58,7 @@ public class ObstacleAvoider {
 		do{
 			currentX = odometer.getX();
 			currentY = odometer.getY();
-			wallFollowerController.processUSData(ultraSonicPoller.getDistance());
+			wallFollowerController.processUSData((int) ultraSonicPoller.getDistance());
 		} while(Math.abs(NavigatorUtility.calculateAngleError(pX - currentX, pY - currentY, odometer.getTheta())*180/Math.PI) > wallFollowingAngleError);
 
 		neckMotor.rotate(neckMotor_OFFSET, false);
